fix(dev-studio): complete registry integration - split client/server, single source of truth
AUDIT FIXES ROUND 2:

1. SPLIT REGISTRY INTO CLIENT/Server:
   - workspace-registry.shared.ts: Client-safe types, STUDIO_WORKSPACES, WORKSPACE_ICONS
   - workspace-registry.ts: Server-only exports with env evaluation
   - Removed 'server-only' from shared file
   - Client imports from workspace-registry.shared.ts only

2. SINGLE SOURCE OF TRUTH:
   - DevStudioUnifiedClient now generates WORKSPACES from STUDIO_WORKSPACES
   - Removed hardcoded WORKSPACES array
   - Removed dead code: _registryIds, _workspaceIds, WORKSPACE_COMPONENT_MAP
   - UI list derived directly from registry: label, icon, id

3. TYPE ALIGNMENT:
   - Workspace type derives from StudioWorkspaceId
   - All render cases use StudioWorkspaceId
   - normalizeWorkspace returns StudioWorkspaceId
   - openWorkspace accepts StudioWorkspaceId

4. PERMISSION CHECKING:
   - Uses getWorkspaceById() from shared registry
   - Both sidebar and MobileTabs use registry

5. CFD DESCRIPTION CORRECTED:
   - Changed from 'OpenFOAM project configuration and execution'
   - To 'Configure and validate OpenFOAM simulation projects'
   - Execution NOT claimed until solver verified

6. LEGACY WORKSPACES:
   - Added command, files, integrations, upload, operations, errors, video
   - All IDs now exist in StudioWorkspaceId type
   - All render with appropriate panels
